requirements.test.py2.txt
atomicwrites==1.1.5
attrs==18.1.0
bitstruct==4.0.0
-e .[test]
coverage==4.5.1
funcsigs==1.0.2
future==0.16.0
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.7.1
py==1.5.4
pytest==4.0.0
scandir==1.9.0
six==1.11.0
tox==3.2.1
typing==3.6.6
virtualenv==16.0.0
xlwt==1.3.0
xlrd==1.1.0
lxml
